+++ MediumDarwin.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from littledarwin import LittleDarwin
+from mediumdarwin import LittleDarwin
 
 if __name__ == "__main__":
     LittleDarwin.main()
