--- littledarwin/__init__.py
@@ -1,11 +1,0 @@
-"""
-__init__ script for littledarwin package
-"""
-
-from littledarwin.LittleDarwin import littleDarwinVersion
-
-__version__ = littleDarwinVersion
-__author__ = "Ali Parsai"
-__package__ = 'littledarwin'
-__license__ = 'GNU GPLv3'
-__url__ = 'https://littledarwin.parsai.net/'
