+++ mediumdarwin/__init__.py
@@ -1,0 +1,11 @@
+"""
+__init__ script for mediumdarwin package
+"""
+
+from mediumdarwin.LittleDarwin import littleDarwinVersion
+
+__version__ = littleDarwinVersion
+__author__ = "Ali Parsai"
+__package__ = 'mediumdarwin'
+__license__ = 'GNU GPLv3'
+__url__ = 'https://mediumdarwin.parsai.net/'
